requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pytesseract>=0.3.10
Pillow>=10.0.0
